Remove WikipediaInfoCollector and replace with WikipediaService
